// src/API.ts
export type Todo = {
  __typename: "Todo",
  id: string,
  name: string,
  description?: string | null,
  completed?: boolean | null,
  createdAt: string,
  updatedAt: string,
};
// ...
export type OnUpdateTodoSubscription = {
  onUpdateTodo?:  {
    __typename: "Todo",
    id: string,
    name: string,
    description?: string | null,
    completed?: boolean | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type OnDeleteTodoSubscriptionVariables = {
  filter?: ModelSubscriptionTodoFilterInput | null,
};

export type OnDeleteTodoSubscription = {
  onDeleteTodo?:  {
    __typename: "Todo",
    id: string,
    name: string,
    description?: string | null,
    completed?: boolean | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};